move config server-fn implementations into server-only modules
The admin routes import getEffectiveConfigFn/setConfigValuesFn and the
instance-config entity fns, so those modules are in the client graph. Their
*Impl functions used @/lib/auth/helpers (getRequestHeaders) outside a
createServerFn handler, so the server-only import survived tree-shaking and
the vite build failed import protection. Move the impls behind .server.ts
siblings referenced only from handler bodies; tests import them directly.
