+++ apps/web/src/server/instance-config.test.ts
@@ -111,7 +111,7 @@ import {
 	listProviderCredentialsImpl,
 	setOrganizationSettingsImpl,
 	upsertProviderCredentialImpl,
-} from "./instance-config";
+} from "./instance-config.server";
 
 beforeEach(() => {
 	state.entityDeny = null;
